+++ services/src/main/java/com/epam/esm/web/impl/TagServiceImpl.java
@@ -1,20 +1,23 @@
 package com.epam.esm.web.impl;
 
+import com.epam.esm.converter.CertificateConverter;
 import com.epam.esm.converter.TagConverter;
+import com.epam.esm.dto.CertificateDto;
 import com.epam.esm.dto.TagDto;
 import com.epam.esm.exception.DuplicateResourceException;
+import com.epam.esm.exception.PaginationException;
 import com.epam.esm.exception.ResourceIsUsedException;
 import com.epam.esm.exception.ResourceNotFoundException;
 import com.epam.esm.model.Tag;
+import com.epam.esm.validator.PageValidator;
 import com.epam.esm.validator.Validator;
 import com.epam.esm.web.TagRepository;
 import com.epam.esm.web.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
-import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -23,23 +26,32 @@ public class TagServiceImpl implements TagService {
 
   private final TagRepository tagRepository;
 
+  private final String CHECK_TAG_EXISTENCE =
+      "select count(certificate_id) from certificateTags where tag_id=?";
+
   private final Validator<TagDto> tagDtoValidator;
+  private final PageValidator pageValidator;
 
   @Autowired
-  public TagServiceImpl(TagRepository tagRepository, Validator<TagDto> validator) {
+  public TagServiceImpl(
+      TagRepository tagRepository, Validator<TagDto> validator, PageValidator pageValidator) {
     this.tagRepository = tagRepository;
     this.tagDtoValidator = validator;
+    this.pageValidator = pageValidator;
   }
 
   @Override
   public TagDto save(TagDto tag) {
     tagDtoValidator.validate(tag);
-    try{
-      Tag newTag = tagRepository.create(TagConverter.convertDtoToModel(tag));
-      tag.setId(newTag.getId());
-    }catch (DataIntegrityViolationException e){
-      throw new DuplicateResourceException("tag with name = " + tag.getName() + " already exists", 40901);
+    List<TagDto> tags = getAll();
+    for (TagDto tagDto : tags) {
+      if (tagDto.getName().equals(tag.getName())) {
+        throw new DuplicateResourceException(
+            "tag with name = " + tag.getName() + " already exists", 40901);
+      }
     }
+    Tag newTag = tagRepository.create(TagConverter.convertDtoToModel(tag));
+    tag.setId(newTag.getId());
     return tag;
   }
 
@@ -53,7 +65,7 @@ public class TagServiceImpl implements TagService {
   @Override
   public TagDto getById(int id) {
     Tag tag = tagRepository.findOne(id);
-    if(tag == null){
+    if (tag == null) {
       throw new ResourceNotFoundException("tag with id = " + id + " does not exist", 40401);
     }
     return TagConverter.convertModelToDto(tag);
@@ -61,14 +73,33 @@ public class TagServiceImpl implements TagService {
 
   @Override
   public int delete(int id) {
-    try{
-      tagRepository.delete(id);
-    }catch (JpaObjectRetrievalFailureException e){
-      throw new ResourceNotFoundException("tag with id = " + id + " does not exist", 40401);
-    }
-    catch (JpaSystemException ex){
-      throw new ResourceIsUsedException("tag with id = " + id + " is in use", 40901);
+    if (!isResourceExist(id)) {
+      throw new ResourceNotFoundException("certificate with id = " + id + " does not exist", 40402);
+    } else {
+      List<Object> params = new ArrayList<>();
+      params.add(id);
+      BigInteger tagCounter =
+          (BigInteger) tagRepository.doNativeGetQuery(CHECK_TAG_EXISTENCE, params);
+      if (tagCounter.intValue() > 0) {
+        throw new ResourceIsUsedException("tag with id = " + id + " is in use", 40901);
+      } else {
+        tagRepository.delete(id);
+      }
     }
     return 1;
+  }
+
+  @Override
+  public List<TagDto> getPaginated(Integer page, Integer size) {
+    pageValidator.validate(page, size);
+    int from = (page - 1) * size;
+    List<TagDto> tags =
+        tagRepository.getPaginated(from, size).stream()
+            .map(TagConverter::convertModelToDto)
+            .collect(Collectors.toList());
+    if (tags.isEmpty()) {
+      throw new PaginationException("this page does not exist", 404);
+    }
+    return tags;
   }
 }
